code rewrite file rerangment
